style: simplification.

+++ file.c
@@ -96,7 +96,7 @@ file_sopen(const char *const fname, const int flags, int *const fd)
 	(void) memset(&how, 0, sizeof(how));
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wsign-conversion"
-	how.flags = ((uint64_t) flags | O_CLOEXEC);
+	how.flags = flags | O_CLOEXEC;
 #pragma GCC diagnostic pop
 	how.resolve = RESOLVE_NO_SYMLINKS | RESOLVE_NO_MAGICLINKS;
